requirements.txt
annotated-types==0.7.0
azure-core==1.35.0
azure-data-tables==12.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.7.0
email-validator==2.2.0
idna==3.10
isodate==0.7.2
multidict==6.6.4
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
requests==2.32.4
six==1.17.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
yarl==1.20.1
